+++ src/tests/roborockService.setSelectedAreas.test.ts
@@ -38,6 +38,7 @@ describe('RoborockService - startClean', () => {
           clearStorageOnStartup: false,
           showRoutinesAsRoom: false,
           includeDockStationStatus: false,
+          includeVacuumErrorStatus: false,
           forceRunAtDefault: false,
           useVacationModeToSendVacuumToDock: false,
           enableCleanModeMapping: false,