+++ src/tests/services/roborockService/roborockService.connectionService.test.ts
@@ -36,6 +36,7 @@ describe('initializeMessageClientForLocal', () => {
           clearStorageOnStartup: false,
           showRoutinesAsRoom: false,
           includeDockStationStatus: false,
+          includeVacuumErrorStatus: false,
           forceRunAtDefault: false,
           useVacationModeToSendVacuumToDock: false,
           enableCleanModeMapping: false,